YYC-109: split src/agent.rs into agent/* submodules
Monolithic agent.rs (2467 lines) was a maintenance hazard. Split
along functional seams:

- mod.rs: Agent struct, AgentBuilder, ModelSelection, init,
  diff_sink/pricing/active_model accessors, replace_history,
  flatten_for_message
- run.rs: run_prompt, run_prompt_stream(_with_cancel),
  prepare_stream_turn, compact_stream_messages_if_needed,
  collect_stream_response, execute_stream_tool_calls,
  sanitize_orphan_tool_messages, empty_terminal_message
- dispatch.rs: dispatch_tool, summarize_tool_args / _result,
  preview_output, elided_lines
- provider.rs: switch_model, switch_provider,
  restore_persisted_provider, fetch_catalog_for,
  resolve_model_selection, available_models
- session.rs: start_session, end_session, resume_session,
  continue_last_session, fork_session, memory accessor
- tests.rs: all unit tests

Cross-module access uses pub(in crate::agent) for the helpers
that span submodules. Behavior unchanged — same 297 tests pass.
